# puptas/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'PUPTAS API Documentation'
  description: 'External integration API for the PUPTAS enrollment system. Provides Student, Program, and Medical endpoints for authorized external partners via Passport client credentials (OAuth2), plus a secure HMAC-signed medical webhook.'
  version: 1.0.0
servers:
  - url: 'http://localhost:8000'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'All external API endpoints require a Passport client-credentials access token sent as a Bearer token. The medical webhook (POST /api/v1/webhooks/medical-result) additionally requires an HMAC `X-Medical-Signature` header (SHA256 of the raw JSON body using the shared `services.medical_webhook.secret`).'
security:
  -
    default: []
paths:
...
  '/api/v1/medical/applicants/idp/{idpUserId}':
    get:
      summary: 'Look up an eligible medical applicant by IDP User ID'
      operationId: lookUpAnEligibleMedicalApplicantByIDPUserID
      description: "Returns identity, contact, education, program, and medical-process status\nfor an applicant who has cleared grade evaluation and interviewing and is\ncurrently pending medical.\n\nRequires the `medical-read` OAuth scope."
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 345
                    idp_user_id: 8f9c6a9b-1a2b-3c4d-5e6f-7a8b9c0d1e2f
                    reference_number: T2026-000123
                    salutation: null
                    firstname: Juan
                    middlename: Santos
                    extension_name: null
                    lastname: 'Dela Cruz'
                    sex: Male
                    email: juan.delacruz@example.com
                    date_graduated: '2025-05-30'
                    strand: STEM
                    track: Academic
                    application:
                      id: 456
                      status: admitted
                      created_at: '2026-06-01T08:00:00.000000Z'
                    program:
                      id: 7
                      code: BSIT
                      name: 'Bachelor of Science in Information Technology'
                    medical_process_status: in_progress
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 345
                      idp_user_id:
                        type: string
                        example: 8f9c6a9b-1a2b-3c4d-5e6f-7a8b9c0d1e2f
                      reference_number:
                        type: string
                        example: T2026-000123
                      salutation:
                        type: string
                        example: null
                        nullable: true
                      firstname:
                        type: string
                        example: Juan
                      middlename:
                        type: string
                        example: Santos
                      extension_name:
                        type: string
                        example: null
                        nullable: true
                      lastname:
                        type: string
                        example: 'Dela Cruz'
                      sex:
                        type: string
                        example: Male
                      email:
                        type: string
                        example: juan.delacruz@example.com
                      date_graduated:
                        type: string
                        example: '2025-05-30'
                      strand:
                        type: string
                        example: STEM
                      track:
                        type: string
                        example: Academic
                      application:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 456
                          status:
                            type: string
                            example: admitted
                          created_at:
                            type: string
                            example: '2026-06-01T08:00:00.000000Z'
                      program:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 7
                          code:
                            type: string
                            example: BSIT
                          name:
                            type: string
                            example: 'Bachelor of Science in Information Technology'
                      medical_process_status:
                        type: string
                        example: in_progress
        404:
          description: 'Applicant not found'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Applicant not found or not eligible for medical yet.'
                properties:
                  message:
                    type: string
                    example: 'Applicant not found or not eligible for medical yet.'
      tags:
        - 'Medical Integration'
    parameters:
      -
        in: path
        name: idpUserId
        description: "The applicant's IDP user UUID."
        example: 8f9c6a9b-1a2b-3c4d-5e6f-7a8b9c0d1e2f
        required: true
        schema:
          type: string
  '/api/v1/medical/applicants/{referenceNumber}':
    get:
      summary: 'Look up an eligible medical applicant by reference number'
      operationId: lookUpAnEligibleMedicalApplicantByReferenceNumber
      description: "Returns identity, contact, education, program, and medical-process status\nfor an applicant who has cleared grade evaluation and interviewing and is\ncurrently pending medical.\n\nRequires the `medical-read` OAuth scope."
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 345
                    idp_user_id: 8f9c6a9b-1a2b-3c4d-5e6f-7a8b9c0d1e2f
                    reference_number: T2026-000123
                    salutation: null
                    firstname: Juan
                    middlename: Santos
                    extension_name: null
                    lastname: 'Dela Cruz'
                    sex: Male
                    email: juan.delacruz@example.com
                    date_graduated: '2025-05-30'
                    strand: STEM
                    track: Academic
                    application:
                      id: 456
                      status: admitted
                      created_at: '2026-06-01T08:00:00.000000Z'
                    program:
                      id: 7
                      code: BSIT
                      name: 'Bachelor of Science in Information Technology'
                    medical_process_status: in_progress
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 345
                      idp_user_id:
                        type: string
                        example: 8f9c6a9b-1a2b-3c4d-5e6f-7a8b9c0d1e2f
                      reference_number:
                        type: string
                        example: T2026-000123
                      salutation:
                        type: string
                        example: null
                        nullable: true
                      firstname:
                        type: string
                        example: Juan
                      middlename:
                        type: string
                        example: Santos
                      extension_name:
                        type: string
                        example: null
                        nullable: true
                      lastname:
                        type: string
                        example: 'Dela Cruz'
                      sex:
                        type: string
                        example: Male
                      email:
                        type: string
                        example: juan.delacruz@example.com
                      date_graduated:
                        type: string
                        example: '2025-05-30'
                      strand:
                        type: string
                        example: STEM
                      track:
                        type: string
                        example: Academic
                      application:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 456
                          status:
                            type: string
                            example: admitted
                          created_at:
                            type: string
                            example: '2026-06-01T08:00:00.000000Z'
                      program:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 7
                          code:
                            type: string
                            example: BSIT
                          name:
                            type: string
                            example: 'Bachelor of Science in Information Technology'
                      medical_process_status:
                        type: string
                        example: in_progress
        404:
          description: 'Applicant not found'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Applicant not found or not eligible for medical yet.'
                properties:
                  message:
                    type: string
                    example: 'Applicant not found or not eligible for medical yet.'
      tags:
        - 'Medical Integration'
    parameters:
      -
        in: path
        name: referenceNumber
        description: "The applicant's application reference number."
        example: T2026-000123
        required: true
        schema:
          type: string
...